添加自定义校验配置项
