+++ server/config/config.go
@@ -186,8 +186,9 @@ func loadServerConfig() ServerConfig {
 			SchemaDisable:  GetBool("registry.schema.disable", false, WithENV("SCHEMA_DISABLE")),
 			SchemaRootPath: GetString("registry.schema.schemaRootPath", "", WithENV("SCHEMA_ROOT_PATH")),
 
-			EnableRBAC:     GetBool("rbac.enable", false, WithStandby("rbac_enabled")),
-			AllowMissToken: GetBool("rbac.allowMissToken", false, WithStandby("rbac_allow_missToken")),
+			EnableRBAC:           GetBool("rbac.enable", false, WithStandby("rbac_enabled")),
+			AllowMissToken:       GetBool("rbac.allowMissToken", false, WithStandby("rbac_allow_missToken")),
+			EnableCustomValidate: GetBool("registry.enableCustomValidate", true, WithStandby("enable_custom_validate")),
 		},
 	}
 }
